2019/12/2

+++ lab2/DebugPractice/DebugExercise1.java
@@ -2,9 +2,10 @@
  * Exercise for learning how the debug, breakpoint, and step-into
  * feature work.
  */
+// Change the int type to float.
 public class DebugExercise1 {
-    public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+    public static int divideThenRound(float top, float bottom) {
+        float quotient = top / bottom;
         int result = Math.round(quotient);
         return result;
     }
